Add elm-lang stuff
This includeds ignoring elm-stuff/ folders and the elm.js file that gets
created. Additionally, it strips whitespace from elm files, and ignores
them inside of nerdtree.

Also, I personally never use Ack and have switched to using rg. Rg
techincally is not a fuzzy finder, but neither is ack. So for now I have
replaced my ctlr-p with rg when it is available, which works well enough
for small projects.

Additionally, this allows \p to toggle paste mode.