Update fzf cmd, md width, and nerdtree ignore
This updates the fzf command to just use `find` as it is plenty fast. It
also ignores the tmp folder and the _site folder which is from jekyll
blogs.

This should probably be cleaned up into maybe logical collections of
things to ignore for a language and combined with nerdtree ignore, but
for now this works.